feat(typescript): add type definitions
Mostly based on the Java code of Besu and the RPC API
documentation + reading through the source code of
web3-eea itself.

Not claiming that the coverage is 100% but definitely
close to it. Happy to add new methods/fix existing ones
if something seems off (I haven't been able to test
all of it out just yet).

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>